backend/requirements.txt
Flask==2.3.3
SQLAlchemy==1.4.52
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
python-dotenv==1.0.1
requests==2.31.0
